requirements.txt
scikit-learn>=1.3.0
numpy>=1.24.0
colorama>=0.4.6
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
